modify database moudule
