+++ AElf.Benchmark/Program.cs
@@ -107,7 +107,7 @@ namespace AElf.Benchmark
                 var builder = new ContainerBuilder();
                 //builder.RegisterModule(new MainModule());
                 builder.RegisterModule(new StateDictatorModule());
-                builder.RegisterModule(new DatabaseModule());
+                builder.RegisterModule(new DatabaseAutofacModule());
                 builder.RegisterModule(new LoggerModule());
                 builder.RegisterModule(new StorageModule());
                 builder.RegisterModule(new ServicesModule());
