+++ AElf.Contracts.SideChain.Tests/ConfigureTestFramework.cs
@@ -1,4 +1,5 @@
 ﻿using AElf.ChainController;
+using AElf.Database;
 using AElf.Kernel;
 using AElf.SmartContract;
 using AElf.Kernel.Modules.AutofacModule;
@@ -38,7 +39,7 @@ namespace AElf.Contracts.SideChain.Tests
             builder.RegisterType(typeof(Hash)).As(typeof(IHash));
             builder.RegisterGeneric(typeof(Serializer<>)).As(typeof(ISerializer<>));
             
-            builder.RegisterModule(new DatabaseModule());
+            builder.RegisterModule(new DatabaseAutofacModule());
             builder.RegisterModule(new LoggerModule());
             builder.RegisterModule(new StorageModule());
             builder.RegisterModule(new ServicesModule());
