+++ AElf.Database/DatabaseAElfModule.cs
@@ -10,18 +10,7 @@ namespace AElf.Database
     {
         public void Init(ContainerBuilder builder)
         {
-            switch (DatabaseConfig.Instance.Type)
-            {
-                case DatabaseType.InMemory:
-                    builder.RegisterType<InMemoryDatabase>().As<IKeyValueDatabase>().SingleInstance();
-                    break;
-                case DatabaseType.Redis:
-                    builder.RegisterType<RedisDatabase>().As<IKeyValueDatabase>().SingleInstance();
-                    break;
-                case DatabaseType.Ssdb:
-                    builder.RegisterType<SsdbDatabase>().As<IKeyValueDatabase>().SingleInstance();
-                    break;
-            }
+            builder.RegisterModule(new DatabaseAutofacModule());
         }
 
         public void Run(ILifetimeScope scope)
