+++ AElf.Kernel/KernelAElfModule.cs
@@ -7,17 +7,7 @@ namespace AElf.Kernel
     {
         public void Init(ContainerBuilder builder)
         {
-            builder.RegisterInstance<IHash>(new Hash()).As<Hash>();
-            
-            var assembly1 = typeof(ISerializer<>).Assembly;
-            builder.RegisterAssemblyTypes(assembly1).AsImplementedInterfaces();
-            
-            var assembly2 = typeof(BlockHeader).Assembly;
-            builder.RegisterAssemblyTypes(assembly2).AsImplementedInterfaces();
-
-            builder.RegisterType(typeof(Hash)).As(typeof(IHash));
-
-            builder.RegisterGeneric(typeof(Serializer<>)).As(typeof(ISerializer<>));
+            builder.RegisterModule(new KernelAutofacModule());
         }
 
         public void Run(ILifetimeScope scope)
