+++ AElf.SmartContract/SmartContractAutofacModule.cs
@@ -1,0 +1,12 @@
+﻿using Autofac;
+
+namespace AElf.SmartContract
+{
+    public class SmartContractAutofacModule : Module
+    {
+        protected override void Load(ContainerBuilder builder)
+        {
+            
+        }
+    }
+}
